Correctly show the buffer status in the player example
QMediaPlayer::bufferProgress() is float based in Qt 6 and return a
number between 0 and 1. Adjust the player example to correctly
convert the numbers.

Pick-to: 6.2

+++ examples/multimediawidgets/player/player.h
@@ -101,7 +101,7 @@ private slots:
 
     void statusChanged(QMediaPlayer::MediaStatus status);
     void stateChanged(QMediaPlayer::PlaybackState state);
-    void bufferingProgress(int progress);
+    void bufferingProgress(float progress);
     void videoAvailableChanged(bool available);
 
     void selectAudioStream();
